+++ pkg/controller/operator/seed/reconciler.go
@@ -26,7 +26,7 @@ import (
 	"k8c.io/kubermatic/v2/pkg/controller/operator/common"
 	"k8c.io/kubermatic/v2/pkg/controller/operator/common/vpa"
 	"k8c.io/kubermatic/v2/pkg/controller/operator/seed/resources/gatekeeper"
-	"k8c.io/kubermatic/v2/pkg/controller/operator/seed/resources/kubermatic"
+	kubermaticseed "k8c.io/kubermatic/v2/pkg/controller/operator/seed/resources/kubermatic"
 	"k8c.io/kubermatic/v2/pkg/controller/operator/seed/resources/nodeportproxy"
 	kubermaticv1 "k8c.io/kubermatic/v2/pkg/crd/kubermatic/v1"
 	operatorv1alpha1 "k8c.io/kubermatic/v2/pkg/crd/operator/v1alpha1"
@@ -185,7 +185,7 @@ func (r *Reconciler) cleanupDeletedSeed(cfg *operatorv1alpha1.KubermaticConfigur
 
 	log.Debug("Seed was deleted, cleaning up cluster-wide resources")
 
-	if err := common.CleanupClusterResource(client, &rbacv1.ClusterRoleBinding{}, kubermatic.ClusterRoleBindingName(cfg)); err != nil {
+	if err := common.CleanupClusterResource(client, &rbacv1.ClusterRoleBinding{}, kubermaticseed.ClusterRoleBindingName(cfg)); err != nil {
 		return fmt.Errorf("failed to clean up ClusterRoleBinding: %v", err)
 	}
 
@@ -201,7 +201,7 @@ func (r *Reconciler) cleanupDeletedSeed(cfg *operatorv1alpha1.KubermaticConfigur
 		return fmt.Errorf("failed to clean up ValidatingWebhookConfiguration: %v", err)
 	}
 
-	if err := common.CleanupClusterResource(client, &admissionregistrationv1.ValidatingWebhookConfiguration{}, common.ClusterAdmissionWebhookName); err != nil {
+	if err := common.CleanupClusterResource(client, &admissionregistrationv1.ValidatingWebhookConfiguration{}, kubermaticseed.ClusterAdmissionWebhookName); err != nil {
 		return fmt.Errorf("failed to clean up Seed ValidatingWebhookConfiguration: %v", err)
 	}
 
@@ -296,7 +296,7 @@ func (r *Reconciler) reconcileServiceAccounts(cfg *operatorv1alpha1.KubermaticCo
 	log.Debug("reconciling Kubermatic ServiceAccounts")
 
 	creators := []reconciling.NamedServiceAccountCreatorGetter{
-		kubermatic.ServiceAccountCreator(cfg, seed),
+		kubermaticseed.ServiceAccountCreator(cfg, seed),
 	}
 
 	if !seed.Spec.NodeportProxy.Disable {
@@ -385,7 +385,7 @@ func (r *Reconciler) reconcileClusterRoleBindings(cfg *operatorv1alpha1.Kubermat
 	log.Debug("reconciling ClusterRoleBindings")
 
 	creators := []reconciling.NamedClusterRoleBindingCreatorGetter{
-		kubermatic.ClusterRoleBindingCreator(cfg, seed),
+		kubermaticseed.ClusterRoleBindingCreator(cfg, seed),
 	}
 
 	if !seed.Spec.NodeportProxy.Disable {
@@ -407,14 +407,14 @@ func (r *Reconciler) reconcileConfigMaps(cfg *operatorv1alpha1.KubermaticConfigu
 	log.Debug("reconciling ConfigMaps")
 
 	creators := []reconciling.NamedConfigMapCreatorGetter{
-		kubermatic.BackupContainersConfigMapCreator(cfg),
+		kubermaticseed.BackupContainersConfigMapCreator(cfg),
 	}
 
-	if creator := kubermatic.ClusterNamespacePrometheusScrapingConfigsConfigMapCreator(cfg); creator != nil {
+	if creator := kubermaticseed.ClusterNamespacePrometheusScrapingConfigsConfigMapCreator(cfg); creator != nil {
 		creators = append(creators, creator)
 	}
 
-	if creator := kubermatic.ClusterNamespacePrometheusRulesConfigMapCreator(cfg); creator != nil {
+	if creator := kubermaticseed.ClusterNamespacePrometheusRulesConfigMapCreator(cfg); creator != nil {
 		creators = append(creators, creator)
 	}
 
@@ -432,8 +432,8 @@ func (r *Reconciler) reconcileSecrets(cfg *operatorv1alpha1.KubermaticConfigurat
 		common.ExtraFilesSecretCreator(cfg),
 		common.SeedWebhookServingCASecretCreator(cfg),
 		common.SeedWebhookServingCertSecretCreator(cfg, client),
-		common.ClusterWebhookServingCASecretCreator(cfg),
-		common.ClusterWebhookServingCertSecretCreator(cfg, client),
+		kubermaticseed.ClusterWebhookServingCASecretCreator(cfg),
+		kubermaticseed.ClusterWebhookServingCertSecretCreator(cfg, client),
 	}
 
 	if cfg.Spec.ImagePullSecret != "" {
@@ -466,7 +466,7 @@ func (r *Reconciler) reconcileDeployments(cfg *operatorv1alpha1.KubermaticConfig
 	log.Debug("reconciling Deployments")
 
 	creators := []reconciling.NamedDeploymentCreatorGetter{
-		kubermatic.SeedControllerManagerDeploymentCreator(r.workerName, r.versions, cfg, seed),
+		kubermaticseed.SeedControllerManagerDeploymentCreator(r.workerName, r.versions, cfg, seed),
 	}
 
 	if !seed.Spec.NodeportProxy.Disable {
@@ -512,7 +512,7 @@ func (r *Reconciler) reconcilePodDisruptionBudgets(cfg *operatorv1alpha1.Kuberma
 	log.Debug("reconciling PodDisruptionBudgets")
 
 	creators := []reconciling.NamedPodDisruptionBudgetCreatorGetter{
-		kubermatic.SeedControllerManagerPDBCreator(cfg),
+		kubermaticseed.SeedControllerManagerPDBCreator(cfg),
 	}
 
 	if !seed.Spec.NodeportProxy.Disable {
@@ -531,7 +531,7 @@ func (r *Reconciler) reconcileServices(cfg *operatorv1alpha1.KubermaticConfigura
 
 	creators := []reconciling.NamedServiceCreatorGetter{
 		common.SeedAdmissionServiceCreator(cfg, client),
-		common.ClusterAdmissionServiceCreator(cfg, client),
+		kubermaticseed.ClusterAdmissionServiceCreator(cfg, client),
 	}
 
 	if err := reconciling.ReconcileServices(r.ctx, creators, cfg.Namespace, client, common.OwnershipModifierFactory(seed, r.scheme)); err != nil {
@@ -570,7 +570,7 @@ func (r *Reconciler) reconcileAdmissionWebhooks(cfg *operatorv1alpha1.Kubermatic
 
 	creators := []reconciling.NamedValidatingWebhookConfigurationCreatorGetter{
 		common.SeedAdmissionWebhookCreator(cfg, client),
-		common.ClusterAdmissionWebhookCreator(cfg, client),
+		kubermaticseed.ClusterAdmissionWebhookCreator(cfg, client),
 	}
 
 	if err := reconciling.ReconcileValidatingWebhookConfigurations(r.ctx, creators, "", client); err != nil {
